refactor: Slightly reduce memory usage and improvisation
Changed several types to appropriate types to reducing time
compilation and the memory usage (depends on operating system)
slightly.

Defined the custom type to represents the cursor position type,
it is called `cpos_t`. Aliased from `unsigned int`, and it might
have different sizes on different platforms (16, 32, or 64 bits),
and it is not always guaranteed to be large enough to represent
the size of the largest object.