refactor: Windows Console API refinements and some refactors
In Windows Console API (`windows.h`) the ENABLE_ECHO_INPUT needs
the ENABLE_LINE_INPUT to be enabled. Thus, we need to enable the
ENABLE_ECHO_INPUT if the echoing is required. Besides that,
now C++ compiler will correctly imports the `stdint.h` header
(for C++11 and above). With some refactor to `__whereis_xy` in
implementation for Windows, the coordinates do not need to be
add by one (for both X and Y), because they both is zero-based
coordinates (starting from 0,0).